FINAL_BACKUP/requirements.txt
akshare
pandas
scikit-learn
psycopg2-binary
pyqlib
